requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-heroku==0.3.1
djangorestframework==3.13.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.1.0
